staging_xcm

Trait IntoVersion

pub trait IntoVersion: Sized {
    // Required method
    fn into_version(self, version: Version) -> Result<Self, ()>;

    // Provided method
    fn into_latest(self) -> Result<Self, ()> { ... }
}

Attempt to convert self into a particular version of itself.

Required Methods

fn into_version(self, version: Version) -> Result<Self, ()>

Consume self and return same value expressed in some particular version of XCM.

Provided Methods

fn into_latest(self) -> Result<Self, ()>

Consume self and return same value expressed the latest version of XCM.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.