staging_xcm::v2

Struct MultiLocation

pub struct MultiLocation {
    pub parents: u8,
    pub interior: Junctions,
}

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

A relative path between state-bearing consensus systems.

A location in a consensus system is defined as an isolatable state machine held within global consensus. The location in question need not have a sophisticated consensus algorithm of its own; a single account within Ethereum, for example, could be considered a location.

A very-much non-exhaustive list of types of location include:

A (normal, layer-1) block chain, e.g. the Bitcoin mainnet or a parachain.

A layer-0 super-chain, e.g. the Polkadot Relay chain.

A layer-2 smart contract, e.g. an ERC-20 on Ethereum.

A logical functional component of a chain, e.g. a single instance of a pallet on a Frame-based Substrate chain.

An account.

A MultiLocation is a relative identifier, meaning that it can only be used to define the relative path between two locations, and cannot generally be used to refer to a location universally. It is comprised of an integer number of parents specifying the number of times to “escape” upwards into the containing consensus system and then a number of junctions, each diving down and specifying some interior portion of state (which may be considered a “sub-consensus” system).

This specific MultiLocation implementation uses a Junctions datatype which is a Rust enum in order to make pattern matching easier. There are occasions where it is important to ensure that a value is strictly an interior location, in those cases, Junctions may be used.

The MultiLocation value of Null simply refers to the interpreting consensus system.

Fields
parents: u8
👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

The number of parent junctions at the beginning of this MultiLocation.

interior: Junctions
👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

The interior (i.e. non-parent) junctions that this MultiLocation contains.

Implementations

impl MultiLocation

pub fn new(parents: u8, junctions: Junctions) -> MultiLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Creates a new MultiLocation with the given number of parents and interior junctions.

pub fn versioned(self) -> VersionedLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Consume self and return the equivalent VersionedLocation value.

pub const fn here() -> MultiLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Creates a new MultiLocation with 0 parents and a Here interior.

The resulting MultiLocation can be interpreted as the “current consensus system”.

pub const fn parent() -> MultiLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Creates a new MultiLocation which evaluates to the parent context.

pub const fn grandparent() -> MultiLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Creates a new MultiLocation which evaluates to the grand parent context.

pub const fn ancestor(parents: u8) -> MultiLocation

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Creates a new MultiLocation with parents and an empty (Here) interior.

pub const fn is_here(&self) -> bool

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Whether the MultiLocation has no parents and has a Here interior.

pub fn interior(&self) -> &Junctions

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return a reference to the interior field.

pub fn interior_mut(&mut self) -> &mut Junctions

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return a mutable reference to the interior field.

pub const fn parent_count(&self) -> u8

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns the number of Parent junctions at the beginning of self.

pub const fn contains_parents_only(&self, count: u8) -> bool

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns boolean indicating whether self contains only the specified amount of parents and no interior junctions.

pub const fn len(&self) -> usize

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns the number of parents and junctions in self.

pub fn first_interior(&self) -> Option<&Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns the first interior junction, or None if the location is empty or contains only parents.

pub fn last(&self) -> Option<&Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns last junction, or None if the location is empty or contains only parents.

pub fn split_first_interior(self) -> (MultiLocation, Option<Junction>)

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Splits off the first interior junction, returning the remaining suffix (first item in tuple) and the first element (second item in tuple) or None if it was empty.

pub fn split_last_interior(self) -> (MultiLocation, Option<Junction>)

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Splits off the last interior junction, returning the remaining prefix (first item in tuple) and the last element (second item in tuple) or None if it was empty or if self only contains parents.

pub fn push_interior(&mut self, new: Junction) -> Result<(), Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Mutates self, suffixing its interior junctions with new. Returns Err with new in case of overflow.

pub fn push_front_interior(&mut self, new: Junction) -> Result<(), Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Mutates self, prefixing its interior junctions with new. Returns Err with new in case of overflow.

pub fn pushed_with_interior( self, new: Junction, ) -> Result<Self, (Self, Junction)>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Consumes self and returns a MultiLocation suffixed with new, or an Err with the original value of self in case of overflow.

pub fn pushed_front_with_interior( self, new: Junction, ) -> Result<Self, (Self, Junction)>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Consumes self and returns a MultiLocation prefixed with new, or an Err with the original value of self in case of overflow.

pub fn at(&self, i: usize) -> Option<&Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns the junction at index i, or None if the location is a parent or if the location does not contain that many elements.

pub fn at_mut(&mut self, i: usize) -> Option<&mut Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns a mutable reference to the junction at index i, or None if the location is a parent or if it doesn’t contain that many elements.

pub fn dec_parent(&mut self)

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Decrements the parent count by 1.

pub fn take_first_interior(&mut self) -> Option<Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Removes the first interior junction from self, returning it (or None if it was empty or if self contains only parents).

pub fn take_last(&mut self) -> Option<Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Removes the last element from interior, returning it (or None if it was empty or if self only contains parents).

pub fn match_and_split(&self, prefix: &MultiLocation) -> Option<&Junction>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Ensures that self has the same number of parents as prefix, its junctions begins with the junctions of prefix and that it has a single Junction item following. If so, returns a reference to this Junction item.

Example

let mut m = MultiLocation::new(1, [PalletInstance(3), OnlyChild].into());
assert_eq!(
    m.match_and_split(&MultiLocation::new(1, [PalletInstance(3)].into())),
    Some(&OnlyChild),
);
assert_eq!(m.match_and_split(&MultiLocation::new(1, Here)), None);

pub fn starts_with(&self, prefix: &MultiLocation) -> bool

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Returns whether self has the same number of parents as prefix and its junctions begins with the junctions of prefix.

Example

let m = MultiLocation::new(1, [PalletInstance(3), OnlyChild, OnlyChild].into());
assert!(m.starts_with(&MultiLocation::new(1, [PalletInstance(3)].into())));
assert!(!m.starts_with(&MultiLocation::new(1, [GeneralIndex(99)].into())));
assert!(!m.starts_with(&MultiLocation::new(0, [PalletInstance(3)].into())));

pub fn append_with(&mut self, suffix: Junctions) -> Result<(), Junctions>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Mutate self so that it is suffixed with suffix.

Does not modify self and returns Err with suffix in case of overflow.

Example

let mut m = MultiLocation::new(1, [Parachain(21)].into());
assert_eq!(m.append_with([PalletInstance(3)].into()), Ok(()));
assert_eq!(m, MultiLocation::new(1, [Parachain(21), PalletInstance(3)].into()));

pub fn prepend_with( &mut self, prefix: MultiLocation, ) -> Result<(), MultiLocation>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Mutate self so that it is prefixed with prefix.

Does not modify self and returns Err with prefix in case of overflow.

Example

let mut m = MultiLocation::new(2, [PalletInstance(3)].into());
assert_eq!(m.prepend_with(MultiLocation::new(1, [Parachain(21), OnlyChild].into())), Ok(()));
assert_eq!(m, MultiLocation::new(1, [PalletInstance(3)].into()));

pub fn reanchored( self, target: &MultiLocation, ancestry: &MultiLocation, ) -> Result<Self, Self>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Consume self and return the value representing the same location from the point of view of target. The context of self is provided as ancestry.

Returns an Err with the unmodified self in the case of error.

pub fn reanchor( &mut self, target: &MultiLocation, ancestry: &MultiLocation, ) -> Result<(), ()>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Mutate self so that it represents the same location from the point of view of target. The context of self is provided as ancestry.

Does not modify self in case of overflow.

pub fn inverted(&self, target: &MultiLocation) -> Result<MultiLocation, ()>

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Treating self as a context, determine how it would be referenced by a target location.

pub fn simplify(&mut self, context: &Junctions)

👎Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Remove any unneeded parents/junctions in self based on the given context it will be interpreted in.