staging_xcm::v2

Struct Xcm

pub struct Xcm<RuntimeCall>(pub Vec<Instruction<RuntimeCall>>);

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

DEPRECATED. Please use XCMv3 or XCMv4 instead.

Tuple Fields
0: Vec<Instruction<RuntimeCall>>
๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Implementations

impl<RuntimeCall> Xcm<RuntimeCall>

pub fn new() -> Self

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Create an empty instance.

pub fn is_empty(&self) -> bool

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return true if no instructions are held in self.

pub fn len(&self) -> usize

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return the number of instructions held in self.

pub fn or_else(self, f: impl FnOnce() -> Self) -> Self

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Consume and either return self if it contains some instructions, or if itโs empty, then instead return the result of f.

pub fn first(&self) -> Option<&Instruction<RuntimeCall>>

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return the first instruction, if any.

pub fn last(&self) -> Option<&Instruction<RuntimeCall>>

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return the last instruction, if any.

pub fn only(&self) -> Option<&Instruction<RuntimeCall>>

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return the only instruction, contained in Self, iff only one exists (None otherwise).

pub fn into_only(self) -> Result<Instruction<RuntimeCall>, Self>

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

Return the only instruction, contained in Self, iff only one exists (returns self otherwise).

impl<RuntimeCall> Xcm<RuntimeCall>

pub fn into<C>(self) -> Xcm<C>

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.

pub fn from<C>(xcm: Xcm<C>) -> Self

๐Deprecated: XCMv2 will be removed once XCMv5 is released. Please use XCMv3 or XCMv4 instead.