Type Alias staging_xcm::v2::InteriorMultiLocation

pub type InteriorMultiLocation = Junctions;

A relative location which is constrained to be an interior location of the context.

See also MultiLocation.

Aliased Type

enum InteriorMultiLocation {
    Here,
    X1(Junction),
    X2(Junction, Junction),
    X3(Junction, Junction, Junction),
    X4(Junction, Junction, Junction, Junction),
    X5(Junction, Junction, Junction, Junction, Junction),
    X6(Junction, Junction, Junction, Junction, Junction, Junction),
    X7(Junction, Junction, Junction, Junction, Junction, Junction, Junction),
    X8(Junction, Junction, Junction, Junction, Junction, Junction, Junction, Junction),
}

Variants

Here

The interpreting consensus system.

X1(Junction)

A relative path comprising 1 junction.

X2(Junction, Junction)

A relative path comprising 2 junctions.

X3(Junction, Junction, Junction)

A relative path comprising 3 junctions.

X4(Junction, Junction, Junction, Junction)

A relative path comprising 4 junctions.

X5(Junction, Junction, Junction, Junction, Junction)

A relative path comprising 5 junctions.

X6(Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 6 junctions.

X7(Junction, Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 7 junctions.

X8(Junction, Junction, Junction, Junction, Junction, Junction, Junction, Junction)

A relative path comprising 8 junctions.

Implementations

impl Junctions

pub const fn into(self) -> MultiLocation

Convert self into a MultiLocation containing 0 parents.

Similar to Into::into, except that this method can be used in a const evaluation context.

pub const fn into_exterior(self, n: u8) -> MultiLocation

Convert self into a MultiLocation containing n parents.

Similar to Self::into, with the added ability to specify the number of parent junctions.

pub fn first(&self) -> Option<&Junction>

Returns first junction, or None if the location is empty.

pub fn last(&self) -> Option<&Junction>

Returns last junction, or None if the location is empty.

pub fn split_first(self) -> (Junctions, Option<Junction>)

Splits off the first junction, returning the remaining suffix (first item in tuple) and the first element (second item in tuple) or None if it was empty.

pub fn split_last(self) -> (Junctions, Option<Junction>)

Splits off the last junction, returning the remaining prefix (first item in tuple) and the last element (second item in tuple) or None if it was empty.

pub fn take_first(&mut self) -> Option<Junction>

Removes the first element from self, returning it (or None if it was empty).

pub fn take_last(&mut self) -> Option<Junction>

Removes the last element from self, returning it (or None if it was empty).

pub fn push(&mut self, new: Junction) -> Result<(), Junction>

Mutates self to be appended with new or returns an Err with new if would overflow.

pub fn push_front(&mut self, new: Junction) -> Result<(), Junction>

Mutates self to be prepended with new or returns an Err with new if would overflow.

pub fn pushed_with(self, new: Junction) -> Result<Self, (Self, Junction)>

Consumes self and returns a Junctions suffixed with new, or an Err with the original value of self and new in case of overflow.

pub fn pushed_front_with(self, new: Junction) -> Result<Self, (Self, Junction)>

Consumes self and returns a Junctions prefixed with new, or an Err with the original value of self and new in case of overflow.

pub const fn len(&self) -> usize

Returns the number of junctions in self.

pub fn at(&self, i: usize) -> Option<&Junction>

Returns the junction at index i, or None if the location doesn’t contain that many elements.

pub fn at_mut(&mut self, i: usize) -> Option<&mut Junction>

Returns a mutable reference to the junction at index i, or None if the location doesn’t contain that many elements.

pub fn iter(&self) -> JunctionsRefIterator<'_>

Returns a reference iterator over the junctions.

pub fn iter_rev(&self) -> impl Iterator + '_

👎Deprecated: Please use iter().rev()

Returns a reference iterator over the junctions in reverse.

pub fn into_iter_rev(self) -> impl Iterator

👎Deprecated: Please use into_iter().rev()

Consumes self and returns an iterator over the junctions in reverse.

pub fn match_and_split(&self, prefix: &Junctions) -> Option<&Junction>

Ensures that self begins with prefix and that it has a single Junction item following. If so, returns a reference to this Junction item.

Example

let mut m = X3(Parachain(2), PalletInstance(3), OnlyChild);
assert_eq!(m.match_and_split(&X2(Parachain(2), PalletInstance(3))), Some(&OnlyChild));
assert_eq!(m.match_and_split(&X1(Parachain(2))), None);

pub fn starts_with(&self, prefix: &Junctions) -> bool

Returns whether self begins with or is equal to prefix.

Example

let mut j = X3(Parachain(2), PalletInstance(3), OnlyChild);
assert!(j.starts_with(&X2(Parachain(2), PalletInstance(3))));
assert!(j.starts_with(&j));
assert!(j.starts_with(&X1(Parachain(2))));
assert!(!j.starts_with(&X1(Parachain(999))));
assert!(!j.starts_with(&X4(Parachain(2), PalletInstance(3), OnlyChild, OnlyChild)));