Type Alias staging_xcm::v4::opaque::Xcm

pub type Xcm = Xcm<()>;

The basic concrete type of Xcm, which doesn’t make any assumptions about the format of a call other than it is pre-encoded.

Aliased Type

struct Xcm(pub Vec<Instruction<()>, Global>);

Fields
0: Vec<Instruction<()>, Global>
Implementations

impl<Call> Xcm<Call>

pub fn new() -> Self

Create an empty instance.

pub fn is_empty(&self) -> bool

Return true if no instructions are held in self.

pub fn len(&self) -> usize

Return the number of instructions held in self.

pub fn inner(&self) -> &[Instruction<Call>]

Return a reference to the inner value.

pub fn inner_mut(&mut self) -> &mut Vec<Instruction<Call>>

Return a mutable reference to the inner value.

pub fn into_inner(self) -> Vec<Instruction<Call>>

Consume and return the inner value.

pub fn iter(&self) -> impl Iterator<Item = &Instruction<Call>>

Return an iterator over references to the items.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Instruction<Call>>

Return an iterator over mutable references to the items.

pub fn into_iter(self) -> impl Iterator<Item = Instruction<Call>>

Consume and return an iterator over the items.

pub fn or_else(self, f: impl FnOnce() -> Self) -> Self

Consume and either return self if it contains some instructions, or if it’s empty, then instead return the result of f.

pub fn first(&self) -> Option<&Instruction<Call>>

Return the first instruction, if any.

pub fn last(&self) -> Option<&Instruction<Call>>

Return the last instruction, if any.

pub fn only(&self) -> Option<&Instruction<Call>>

Return the only instruction, contained in Self, iff only one exists (None otherwise).

pub fn into_only(self) -> Result<Instruction<Call>, Self>

Return the only instruction, contained in Self, iff only one exists (returns self otherwise).

impl<Call> Xcm<Call>

pub fn builder() -> XcmBuilder<Call, PaymentRequired>

pub fn builder_unpaid() -> XcmBuilder<Call, ExplicitUnpaidRequired>

pub fn builder_unsafe() -> XcmBuilder<Call, AnythingGoes>

impl<Call> Xcm<Call>

pub fn into<C>(self) -> Xcm<C>

pub fn from<C>(xcm: Xcm<C>) -> Self