staging_xcm
Module v5

Version 5 of the Cross-Consensus Message format data structures.

Re-exports

pub use super::v3::GetWeight;

pub use super::v4::MaxDispatchErrorLen;

pub use super::v4::MaybeErrorCode;

pub use super::v4::OriginKind;

pub use super::v4::WeightLimit;

Modules

opaque

prelude

A prelude for importing all types typically used when interacting with XCM messages.

Structs

Ancestor

A unit struct which can be converted into a Location of the inner parents value.

AncestorThen

A unit struct which can be converted into a Location of the inner parents value and the inner interior.

Asset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

AssetId

Location to identify an asset.

Assets

A Vec of Assets.

HintNumVariants

Location

A relative path between state-bearing consensus systems.

MaxPalletNameLen

MaxPalletsInfo

PalletInfo

Parent

A unit struct which can be converted into a Location of parents value 1.

ParentThen

A tuple struct which can be converted into a Location of parents value 1 with the inner interior.

QueryResponseInfo

Information regarding the composition of a query response.

Weight

Xcm

XcmBuilder

Type used to build XCM programs

XcmContext

Contextual data pertaining to a specific list of XCM instructions.

Enums

AnythingGoes

Access to all the instructions

AssetFilter

Asset collection, defined either by a number of Assets or a single wildcard.

AssetInstance

A general identifier for an instance of a non-fungible asset class.

AssetTransferFilter

Matches assets based on inner AssetFilter and tags them for a specific type of asset transfer. Please note: the transfer type is specific to each particular (asset, source, dest) combination, so it should always be built in the context of source after knowing dest.

BodyId

An identifier of a pluralistic body.

BodyPart

A part of a pluralistic body.

Error

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

ExplicitUnpaidRequired

Need to explicitly state it won’t pay for fees

Fungibility

Classification of whether an asset is fungible or not, along with a mandatory amount or instance.

Hint

Instruction

Cross-Consensus Message: A message from one consensus system to another.

Junction

A single item in a path to describe the relative location of a consensus system.

Junctions

Non-parent junctions that can be constructed, up to the length of 8. This specific Junctions implementation uses a Rust enum in order to make pattern matching easier.

LoadedHolding

The holding register was loaded, now to buy execution

NetworkId

A global identifier of a data structure existing within consensus.

Outcome

Outcome of an XCM execution.

PaymentRequired

You need to pay for execution

Response

Response data to a query.

SendError

Error result value when attempting to send an XCM message.

WildAsset

A wildcard representing a set of assets.

WildFungibility

Classification of whether an asset is fungible or not.

Constants

MAX_INSTRUCTIONS_TO_DECODE

MAX_ITEMS_IN_ASSETS

Maximum number of items we expect in a single Assets value. Note this is not (yet) enforced, and just serves to provide a sensible max_encoded_len for Assets.

ROCOCO_GENESIS_HASH

The genesis hash of the Rococo testnet. Used to identify it.

VERSION

WESTEND_GENESIS_HASH

The genesis hash of the Westend testnet. Used to identify it.

Traits

ExecuteXcm

Type of XCM message executor.

PreparedMessage

Reanchorable

SendXcm

Utility for sending an XCM message to a given location.

XcmBuilderState

A trait for types that track state inside the XcmBuilder

XcmWeightInfo

Functions

send_xcm

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

validate_send

Convenience function for using a SendXcm implementation. Just interprets the dest and wraps both in Some before passing them as mutable references into T::send_xcm.

Type Aliases

InteriorLocation

A relative location which is constrained to be an interior location of the context.

QueryId

An identifier for a query.

Result

SendResult

Result value when attempting to send an XCM message.

XcmHash

A hash type for identifying messages.