staging_xcm::v5

Trait ExecuteXcm

pub trait ExecuteXcm<Call> {
    type Prepared: PreparedMessage;

    // Required methods
    fn prepare(
        message: Xcm<Call>,
        weight_limit: Weight,
    ) -> Result<Self::Prepared, InstructionError>;
    fn execute(
        origin: impl Into<Location>,
        pre: Self::Prepared,
        id: &mut XcmHash,
        weight_credit: Weight,
    ) -> Outcome;
    fn charge_fees(location: impl Into<Location>, fees: Assets) -> Result;

    // Provided method
    fn prepare_and_execute(
        origin: impl Into<Location>,
        message: Xcm<Call>,
        id: &mut XcmHash,
        weight_limit: Weight,
        weight_credit: Weight,
    ) -> Outcome { ... }
}

Type of XCM message executor.

Required Associated Types

type Prepared: PreparedMessage

Required Methods

fn prepare( message: Xcm<Call>, weight_limit: Weight, ) -> Result<Self::Prepared, InstructionError>

If it fails, returns the index of the problematic instruction.

fn execute( origin: impl Into<Location>, pre: Self::Prepared, id: &mut XcmHash, weight_credit: Weight, ) -> Outcome

fn charge_fees(location: impl Into<Location>, fees: Assets) -> Result

Deduct some fees to the sovereign account of the given location and place them as per the convention for fees.

Provided Methods

fn prepare_and_execute( origin: impl Into<Location>, message: Xcm<Call>, id: &mut XcmHash, weight_limit: Weight, weight_credit: Weight, ) -> Outcome

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.