Crate staging_xcm

Cross-Consensus Message format data structures.

Modules

latest

lts

opaque

prelude

v2Deprecated

XCM Version 2

v3

Version 3 of the Cross-Consensus Message format data structures.

v4

Version 4 of the Cross-Consensus Message format data structures.

Structs

AlwaysV2

WrapVersion implementation which attempts to always convert the XCM to version 2 before wrapping it.

AlwaysV3

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

AlwaysV4

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

Enums

Unsupported

VersionedAsset

A single Asset value, together with its version code.

VersionedAssetId

A single version’s AssetId value, together with its version code.

VersionedAssets

A single MultiAssets value, together with its version code.

VersionedInteriorLocation

A single InteriorLocation value, together with its version code.

VersionedJunction

A single Junction value, together with its version code.

VersionedLocation

A single Location value, together with its version code.

VersionedNetworkId

A single NetworkId value, together with its version code.

VersionedResponse

A single version’s Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

Constants

MAX_INSTRUCTIONS_TO_DECODE

The maximal number of instructions in an XCM before decoding fails.

MAX_XCM_DECODE_DEPTH

Maximum nesting level for XCM decoding.

Traits

GetVersion

Check and return the Version that should be used for the Xcm datum for the destination Location, which will interpret it.

IdentifyVersion

Used to get the version out of a versioned type.

IntoVersion

Attempt to convert self into a particular version of itself.

TryAs

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination Location which will interpret it.

Type Aliases

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysLts

WrapVersion implementation which attempts to always convert the XCM to the most recent Long- Term-Support version before wrapping it.

Version

A version of XCM.

VersionedInteriorMultiLocationDeprecated

VersionedMultiAssetDeprecated

VersionedMultiAssetsDeprecated

VersionedMultiLocationDeprecated