Crate staging_xcm

Cross-Consensus Message format data structures.

Modules

latest

lts

opaque

prelude

v2

XCM Version 2

v3

Version 3 of the Cross-Consensus Message format data structures.

Structs

AlwaysV2

WrapVersion implementation which attempts to always convert the XCM to version 2 before wrapping it.

AlwaysV3

WrapVersion implementation which attempts to always convert the XCM to version 3 before wrapping it.

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

Enums

Unsupported

VersionedAssetId

A single version’s Response value, together with its version code.

VersionedInteriorMultiLocation

A single InteriorMultiLocation value, together with its version code.

VersionedJunction

A single Junction value, together with its version code.

VersionedMultiAsset

A single MultiAsset value, together with its version code.

VersionedMultiAssets

A single MultiAssets value, together with its version code.

VersionedMultiLocation

A single MultiLocation value, together with its version code.

VersionedNetworkId

A single NetworkId value, together with its version code.

VersionedResponse

A single version’s Response value, together with its version code.

VersionedXcm

A single XCM message, together with its version code.

Constants

MAX_XCM_DECODE_DEPTH

Maximum nesting level for XCM decoding.

Traits

IntoVersion

Attempt to convert self into a particular version of itself.

TryAs

WrapVersion

Convert an Xcm datum into a VersionedXcm, based on a destination MultiLocation which will interpret it.

Type Definitions

AlwaysLatest

WrapVersion implementation which attempts to always convert the XCM to the latest version before wrapping it.

AlwaysLts

WrapVersion implementation which attempts to always convert the XCM to the most recent Long- Term-Support version before wrapping it.

Version

A version of XCM.