staging_xcm

Trait GetVersion

pub trait GetVersion {
    // Required method
    fn get_version_for(dest: &Location) -> Option<Version>;
}

Check and return the Version that should be used for the Xcm datum for the destination Location, which will interpret it.

Required Methods

fn get_version_for(dest: &Location) -> Option<Version>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.