Trait staging_xcm::GetVersion

pub trait GetVersion {
    // Required method
    fn get_version_for(dest: &Location) -> Option<Version>;
}

Check and return the Version that should be used for the Xcm datum for the destination Location, which will interpret it.

Required Methods

fn get_version_for(dest: &Location) -> Option<Version>

Object Safety

This trait is not object safe.